Updating files and packages
