+++ src/data/jmetal/problems/DoubleConfiguration.java
@@ -1,4 +1,4 @@
-package JMetal;
+package data.jmetal.problems;
 
 import java.io.IOException;
 import java.util.ArrayList;
