+++ src/data/jmetal/problems/IntegerProblem.java
@@ -1,4 +1,4 @@
-package JMetal;
+package data.jmetal.problems;
 
 public class IntegerProblem {
