+++ src/data/utils/VariableType.java
@@ -1,4 +1,4 @@
-package utils;
+package data.utils;
 
 /**
  *
